utils/se_inference.py
"""
Simplified SE Inference Module

This is a simplified version of the SE inference module for the SE+ST combined model.
In a real implementation, you would need to integrate with the actual SE model.
"""

import logging
import torch
import torch.nn as nn
from typing import Optional

logger = logging.getLogger(__name__)


class SEInference:
    """
    Simplified SE Inference class for SE+ST combined model.
    
    Note: This is a placeholder implementation. In practice, you would need to:
    1. Install the actual SE model package
    2. Implement proper SE model loading
    3. Handle the actual SE model inference
    """
    
    def __init__(self, input_dim: int = 2000, output_dim: int = 512):
        self.model = None
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.input_dim = input_dim
        self.output_dim = output_dim
    
    def load_model(self, checkpoint_path: str):
        """
        Load SE model from checkpoint.
        
        Args:
            checkpoint_path: Path to SE model checkpoint
        """
        try:
            # This is a placeholder - you would need to implement actual SE model loading
            logger.warning("SE model loading is not implemented in this simplified version.")
            logger.warning("Please install the actual SE model package and implement proper loading.")
            
            # Create a dummy model for demonstration with correct dimensions
            self.model = DummySEModel(input_dim=self.input_dim, output_dim=self.output_dim)
            logger.info(f"Dummy SE model loaded (checkpoint: {checkpoint_path}, input_dim={self.input_dim}, output_dim={self.output_dim})")
            
        except Exception as e:
            logger.error(f"Failed to load SE model: {e}")
            raise
    
    def transform(self, input_path: str, output_path: str, batch_size: int = 1000):
        """
        Transform data using SE model.
        
        Args:
            input_path: Path to input data
            output_path: Path to save transformed data
            batch_size: Batch size for processing
        """
        logger.warning("SE model transformation is not implemented in this simplified version.")
        logger.warning("Please implement actual SE model transformation.")


class DummySEModel(nn.Module):
    """
    Dummy SE model for demonstration purposes.
    
    In practice, this would be replaced with the actual SE model.
    """
    
    def __init__(self, input_dim: int = 2000, output_dim: int = 512):
        super().__init__()
        self.input_dim = input_dim
        self.output_dim = output_dim
        
        # Simple linear transformation as placeholder
        self.encoder = nn.Sequential(
            nn.Linear(input_dim, 1024),
            nn.ReLU(),
            nn.Linear(1024, output_dim),
            nn.ReLU()
        )
    
    def encode_cells(self, cell_expressions: torch.Tensor) -> torch.Tensor:
        """
        Encode cell expressions to state embeddings.
        
        Args:
            cell_expressions: Cell expressions [N_cells, N_genes]
            
        Returns:
            state_embeddings: State embeddings [N_cells, state_dim]
        """
        return self.encoder(cell_expressions)
    
    def forward(self, x):
        return self.encode_cells(x)


# Alternative: If you have access to the actual SE model, you can use this:
"""
# Uncomment and modify this section if you have access to the actual SE model

try:
    from state.emb.inference import Inference as ActualSEInference
    
    class SEInference(ActualSEInference):
        def __init__(self):
            super().__init__()
            
        def load_model(self, checkpoint_path: str):
            super().load_model(checkpoint_path)
            
        def transform(self, input_path: str, output_path: str, batch_size: int = 1000):
            super().transform(input_path, output_path, batch_size)
            
except ImportError:
    logger.warning("Actual SE model not available, using dummy implementation")
"""
