requirements.txt
# Core dependencies
torch>=2.0.0
lightning>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# PyTorch Geometric (GNN)
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Data processing
anndata>=0.9.0
scanpy>=1.9.0
pandas>=1.5.0
numpy>=1.23.0

# Network processing
requests>=2.28.0
geomloss>=0.2.6

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
